producer_service/pyproject.toml
[tool.poetry]
name = "producer-service"
version = "0.1.0"
description = ""
authors = ["goaho7 <goaho7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
kafka-python = "^2.0.2"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
